Update StdEntropyDecoder.java
setmask = (3<<bp)>>1;  works great for unsigned integers however setmask is a signed integer so when bp = 30  the result is negative (-536870912) and thus incorrect. The correct mask result for bit plane 30 should be 1610612736, which is correctly returned by 3<<(bp-1).  This bug results in black 'holes' / 'stripes' in the image as shown here on StackOverflow (https://stackoverflow.com/questions/41977536/black-stain-when-extracting-page-to-image-on-pdfbox-2-0-4)