Disable DO_TIMING blocks for OpenJDK 11 support
The code will not compile as long deprecated
System.runFinalizersOnExit
was removed by JDK-8198249 in Open JDK 11.

As using this timing requires code
modification and was never used "live" in this patch I added a /*
UNCOMMENT  */ comment block around the offending code - it can be
enabled if compiling on Open JDK 8.

Based on #15 by @rleigh-codelibre